restructure fixture for xai contract set deployment add forwarder test
